+++ pyhf/constraints.py
@@ -1,38 +1,5 @@
 from . import get_backend, default_backend
 
-
-class param_set(object):
-    def __init__(self, n_parameters, inits, bounds):
-        self.n_parameters = n_parameters
-        self.suggested_init = inits
-        self.suggested_bounds = bounds
-
-class standard_gaussian_constraint(param_set):
-    def __init__(self, n_parameters, inits, bounds, auxdata):
-        super(standard_gaussian_constraint,self).__init__(n_parameters, inits, bounds)
-        self.pdf_type = 'normal'
-        self.auxdata = auxdata
-
-    def alphas(self, pars):
-        return pars  # the nuisance parameters correspond directly to the alpha
-
-    def expected_data(self, pars):
-        return self.alphas(pars)
-
-class factor_poisson_constraint(param_set):
-    def __init__(self, n_parameters, inits, bounds, auxdata, factors):
-        super(factor_poisson_constraint,self).__init__(n_parameters, inits, bounds)
-        self.pdf_type = 'poisson'
-        self.auxdata = auxdata
-        self.factors = factors
-
-    def alphas(self, pars):
-        tensorlib, _ = get_backend()
-        return tensorlib.product(tensorlib.stack([pars, tensorlib.astensor(self.factors)]), axis=0)
-    def expected_data(self, pars):
-        return self.alphas(pars)
-
-
 class gaussian_constraint_combined(object):
     def __init__(self,pdfconfig):
         # iterate over all constraints order doesn't matter....
