+++ pyhf/modifiers/histosys.py
@@ -2,7 +2,7 @@ import logging
 log = logging.getLogger(__name__)
 
 from . import modifier
-from ..constraints import standard_gaussian_constraint
+from ..paramsets import standard_gaussian_constraint
 
 from .. import get_backend
 from ..interpolate import interpolator
